client/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
requests>=2.31.0
pydantic>=2.6.0
python-dotenv>=1.0.1
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.1.0
langsmith>=0.2.0
google-genai>=1.0.0
